pyproject.toml
[project]
name = "kan-myst-interactive"
version = "0.1.0"
description = "MyST article + Dash dashboard for the KAN paper."
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  # Web layer
  "dash>=2.16,<3.0",
  "plotly>=5.18",
  "gunicorn>=21.2",
  # pykan + the same numeric pins kan-fulltext-demo uses
  "pykan",
  "torch==2.2.2",
  "numpy==1.24.4",
  "matplotlib==3.6.2",
  "pandas==2.0.1",
  "scikit-learn==1.1.3",
  "seaborn",
  "sympy==1.11.1",
  "tqdm==4.66.2",
  "setuptools==65.5.0",
  "pyyaml",
]

[tool.uv]
package = false
